Remove the assignment of TERM
Apparently that's the wrong way to set the default terminal.
I think I knew that, but then again it's 2:40 in the morning.

My excuse: I used to alias `term' as the correct one.

The 'default terminal' exists only in the mind of the window manager.
So the correct place to change that is in my WM's run-control.