Added ZSH configuration files.
I have not yet changed the shell, but if I do, it is likely possible to
configure (via environment variables) so that most of the garbage that
I'm supposed to put in my home directory is instead looked for here.
ZSH strikes me as being configurable.

Which is great, but also shitty, because it's yet another thing to mess
around with instead of actually getting real work done.

"It'll make me more productive, I swear!" =)