Added ALSA configuration file.
Woe betide me the next time an update rolls around.
I hate to say it, but I have little idea what's actually in this file.
It's something about using the pulseaudio server, on a different card.

I should reference the card by name, not number.  This is something to
check in the case of sound problems occuring inexplicably on reboot.

Software mixing now works properly.